Add 64-bit ranges for significance
We want more precision than we currently have so we can set a minimum
significance of 10e-100

+++ src/data_structures/regeffects.rs
@@ -16,7 +16,7 @@ pub struct ObservationData {
     // categorical facet values
     pub facet_ids: Vec<DbID>,
     pub effect_size: f32,
-    pub significance: f32,
+    pub significance: f64,
 }
 
 #[derive(Serialize, Deserialize, Clone, Debug)]
